backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-multipart>=0.0.6
aiohttp>=3.9.1
aiofiles>=23.2.1
pillow>=10.4.0
streamdeck>=0.9.5
websockets>=12.0
python-dotenv>=1.0.0
psutil>=5.9.0
soco>=0.30.0
spotipy>=2.23.0
